# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Management API"
  description: "This is the API documentation for the Mechanic Management system."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header



paths: 
#CUSTOMER ENDPOINTS
  /customers/login:
    post:
      tags:
        - "Customers"
      summary: "Login a customer"
      description: "This endpoint allows a customer to log in."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "kjlasdkfjadslkj"
              message: "Login successful"
              status: "success"
        400:
          description: "Invalid credentials"
  /customers: #endpoint
    post:
      tags:
        - Customers
      summary: "Create a customer"
      description: "Endpoint creates a new customer"
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to create a customer"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              phone: "1234567890"
              email: jdoe@gmail.com
              password: mypassword

    get:
      tags:
        - Customers
      summary: "Get all customers"
      description: "Endpoint returns all customers"
      responses:
        200:
          description: "Retrieved customers successfully"
          schema:
            $ref: "#/definitions/AllCustomers"

    put:
      tags:
        - Customers
      summary: "Update a customer"
      description: "Endpoint updates a customer"
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to update a customer"
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              phone: "1234567890"
              email: jdoe@gmail.com
              password: mypassword
               #endpoint
    delete:
      tags:
        - Customers
      summary: "Delete a customer"
      description: "Endpoint deletes a customer"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer deleted"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "succesfully deleted <customer_id>"
#SERVICE TICKET ENDPOINTS
  /service_tickets:
    post:
      tags:
        - Service Tickets
      summary: "Create a service ticket"
      description: "Endpoint creates a new service ticket"
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to create a service ticket"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          description: "Service ticket created"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
          examples:
            application/json:
              id: 1
              customer_id: 1
              VIN: "1234567890"
              service_date: "2021-01-01"
              service_desc: "Oil change"
              mechanic_ids: [1, 2, 3]
    get:
      tags:
        - Service Tickets
      summary: "Get all service tickets"
      description: "Endpoint returns all service tickets"
      responses:
        200:
          description: "Retrieved service tickets successfully"
          schema:
            $ref: "#/definitions/AllServiceTickets"
  /service_tickets/{id}/edit:
    put:
      tags:
        - Service Tickets
      summary: "Update mechanics on a service ticket"
      description: "Endpoint updates a service ticket"
      parameters:
        - in: "path"
          name: "id"
          description: "Service ticket ID"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Information needed to update a service ticket"
          required: true
          schema:
            $ref: "#/definitions/UpdateServiceTicketPayload"
      responses:
        200:
          description: "Service ticket updated"
          schema:
            $ref: "#/definitions/UpdateServiceTicketResponse"
          examples:
            application/json:
              id: 1
              customer_id: 1
              VIN: "1234567890"
              service_date: "2021-01-01"
              service_desc: "Oil change"
  /service_tickets/{id}:
    put:
      tags:
        - Service Tickets
      summary: "Updates a service ticket"
      description: "Endpoint updates a service ticket"
      parameters:
        - in: "path"
          name: "id"
          description: "Service ticket ID"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Information needed to update a service ticket"
          required: true
          schema:
            $ref: "#/definitions/UpdateServiceTicketMechPayload"
      responses:
        200:
          description: "Service ticket updated"
          schema:
            $ref: "#/definitions/UpdateServiceTicketMechResponse"
          examples:
            application/json:
              add_mechanic_ids: [1, 2, 3]
              remove_mechanic_ids: [4, 5, 6]
              add_part_ids: [1, 2, 3]
    delete:
      tags:
        - Service Tickets
      summary: "Delete a service ticket"
      description: "Endpoint deletes a service ticket"
      parameters:
        - in: "path"
          name: "id"
          description: "Service ticket ID"
          required: true
          type: "integer"
          schema: 
            $ref: "#/definitions/DeleteServiceTicketPayload"
      responses:
        200:
          description: "Service ticket deleted"
          schema:
            $ref: "#/definitions/DeleteServiceTicketResponse"
          examples:
            application/json:
              message: "succesfully deleted <service_ticket_id>"



#MECHANIC ENDPOINTS            
definitions: 
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
    
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"

  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      
    required:
      - name
      - email
      - phone
      - password


  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"

  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        password:
          type: "string"
    
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
    - name
    - email
    - phone
    - password

  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"

  CreateServiceTicketPayload:
    type: "object"
    properties:
      VIN:
        type: "string"
      customer_id:
        type: "integer"
      service_date:
        type: "string"
      service_desc:
        type: "string"
      mechanic_ids:
        type: "array"
        items:
          type: "integer"
    required:
      - VIN
      - service_date
      - service_desc
      - customer_id
      - mechanic_ids
  CreateServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      customer_id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
      service_desc:
        type: "string"
      mechanic_ids:
        type: "array"
        items:
          type: "integer"

  UpdateServiceTicketPayload:
    type: "object"
    properties:
      customer_id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
      service_desc:
        type: "string"
  UpdateServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      customer_id:
        type: "integer"
      service_date:
        type: "string"
      service_desc:
        type: "string"
    required:
      - id
      - VIN
      - service_date
      - service_desc
      - customer_id
  UpdateServiceTicketMechPayload:
    type: "object"
    properties:
      add_mechanic_ids:
        type: "array"
        items:
          type: "integer"
      remove_mechanic_ids:
        type: "array"
        items:
          type: "integer"
      add_part_ids:
        type: "array"
        items:
          type: "integer"
    required:
      - add_mechanic_ids
      - remove_mechanic_ids
      - add_part_ids
  UpdateServiceTicketMechResponse:
    type: "object"
    properties:
      add_mechanic_ids:
        type: "array"
        items:
          type: "integer"
      remove_mechanic_ids:
        type: "array"
        items:
          type: "integer"
      add_part_ids:
        type: "array"
        items:
          type: "integer"
    required:
      - add_mechanic_ids
      - remove_mechanic_ids
      - add_part_ids

  DeleteServiceTicketPayload:
    type: "object"
    properties:
      id:
        type: "integer"
    required:
      - id    
  DeleteServiceTicketResponse:
    type: "object"
    properties:
      id: 
        type:
          "integer"
      message:
        type: "string"